% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{DESEZ_import_data_points}
\alias{DESEZ_import_data_points}
\title{Insert data points from DESEZONIRANJE}
\usage{
DESEZ_import_data_points(
  con,
  config = desezoniranje_config,
  schema = "platform"
)
}
\arguments{
\item{con}{Database connection}

\item{config}{List of configurations. Defaults to desezoniranje_config.}

\item{schema}{Schema name, defaults to "platform"}
}
\value{
Insertion results (invisibly)
}
\description{
Function to prepare and insert DESEZONIRANJE data points. The function first prepares
the required vintages and inserts them, then prepares the data points
table and inserts it. The function returns the results invisibly.
}
\details{
This is a DESEZONIRANJE specific function, which should be followed by the generic
UMARimportR function to write the vintage hashes and clean up redundant
vintages.
}
