% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_relationship_table}
\alias{prepare_category_relationship_table}
\title{Prepare table to insert into `category_relationship` table}
\usage{
prepare_category_relationship_table(
  con,
  config = desezoniranje_config,
  schema = "platform"
)
}
\arguments{
\item{con}{connection to the database}

\item{config}{List of configurations. Defaults to `desezoniranje_config`.}

\item{schema}{schema name}
}
\value{
a dataframe with the `id`, `parent_id`, `source_id` for each relationship
betweeh categories
}
\description{
Helper function that manually prepares the category_relationship table.
Returns table ready to insert into the `category_relationship` table with the db_writing family
of functions from `UMARimportR`.
}
