% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_table}
\alias{prepare_category_table}
\title{Prepare table to insert into `category` table}
\usage{
prepare_category_table(con, config = desezoniranje_config, schema = "platform")
}
\arguments{
\item{con}{connection to the database}

\item{config}{List of configurations. Defaults to `desezoniranje_config`.}

\item{schema}{schema name}
}
\value{
a dataframe with the `id`, `name`, `source_id` for each all the categories
used to describe the DESEZONIRANJE ecosystem
}
\description{
Helper function that manually prepares the category table with field ids and
their names. Returns table ready to insert into the `category` table with the db_writing family
of functions from `UMARimportR`.
}
