% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_series_levels_table}
\alias{prepare_series_levels_table}
\title{Prepare table to insert into `series_levels` table}
\usage{
prepare_series_levels_table(
  con,
  config = desezoniranje_config,
  schema = "platform"
)
}
\arguments{
\item{con}{connection to the database}

\item{config}{List of configurations. Defaults to `desezoniranje_config`.}

\item{schema}{schema name}
}
\value{
a dataframe with the `series_id`, `tab_dim_id`, `value` columns
all the series-level combinatins for this table.
}
\description{
Helper function that extracts the individual levels for each series and
gets the correct dimension id for each one and the correct series id to
keep with the constraints.
Returns table ready to insert into the `series_levels`table with the
db_writing family of functions.
}
