% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_series_table}
\alias{prepare_series_table}
\title{Prepare table to insert into `series` table}
\usage{
prepare_series_table(con, config = desezoniranje_config, schema = "platform")
}
\arguments{
\item{con}{Connection to the database}

\item{config}{List of configurations. Defaults to `desezoniranje_config`.}

\item{schema}{Schema name, defaults to "platform"}
}
\value{
A dataframe with columns: `table_id`, `name_long`, `code`,
  `unit_id`, and `interval_id` for each series.
}
\description{
Prepares series metadata for all tables in desezoniranje_config.
Creates one series for each unique table_id and column_code combination
(excluding period_id).
}
