% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_table_table}
\alias{prepare_table_table}
\title{Prepare table to insert into `table` table}
\usage{
prepare_table_table(
  con,
  config = desezoniranje_config,
  schema = "platform",
  keep_vintage = FALSE
)
}
\arguments{
\item{con}{Connection to the database}

\item{config}{List of configurations. Defaults to `desezoniranje_config`.}

\item{schema}{Schema name, defaults to "platform"}

\item{keep_vintage}{Logical indicating whether to keep vintages, defaults to FALSE}
}
\value{
A list of dataframes, one for each unique table_id in the config.
  Each dataframe contains `code`, `name`, `source_id`, `url`, and `notes` columns.
}
\description{
Helper function that prepares the table metadata from desezoniranje_config.
Returns a list of tables ready to insert into the `table` table with the
db_writing family of functions from `UMARimportR`.
}
