% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_vintage_table}
\alias{prepare_vintage_table}
\title{Prepare table to insert into `vintage` table}
\usage{
prepare_vintage_table(
  file_paths_df,
  con,
  config = desezoniranje_config,
  schema = "platform"
)
}
\arguments{
\item{file_paths_df}{Data frame with source_name, file_path, and file_mtime columns,
typically from `get_all_recent_files()`}

\item{con}{Connection to the database}

\item{config}{List of configurations. Defaults to `desezoniranje_config`.}

\item{schema}{Schema to use for the connection, default is "platform"}
}
\value{
A dataframe with `series_id` and `published` columns
  for all series across all tables.
}
\description{
Helper function that populates the vintage table with the new vintages. It gets
the series ids from the database and adds the publication date from the file
modification time.
}
\details{
Returns table ready to insert into the `vintage` table with the
UMARimportR::insert family of functions.
}
