% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_dimension_levels_table}
\alias{prepare_dimension_levels_table}
\title{Prepare table to insert into `dimension_levels` table}
\usage{
prepare_dimension_levels_table(
  con,
  config = desezoniranje_config,
  schema = "platform"
)
}
\arguments{
\item{con}{Connection to the database}

\item{config}{List of configurations. Defaults to `desezoniranje_config`.}

\item{schema}{Schema name, defaults to "platform"}
}
\value{
A list of dataframes, one for each unique table_id. Each dataframe
  contains `tab_dim_id`, `level_value`, and `level_text` columns.
}
\description{
Helper function that manually prepares the dimension_levels for each
table and gets their codes and text from desezoniranje_config.
Returns a list of tables ready to insert into the `dimension_levels` table
with the db_writing family of functions.
}
